RAG_demo/requirements.txt
requests
faiss-cpu
numpy